gokul_negation/requirements.txt
ipykernel
pandas
groq
python-dotenv
ollama